requirements.txt
# ==============================
# AI ENGINE
# ==============================

fastapi
uvicorn
pydantic
python-dotenv

# ==============================
# LLM INTEGRATION
# ==============================

groq
requests

# ==============================
# AUTOMATION
# ==============================

playwright

# ==============================
# COMPUTER VISION
# ==============================

opencv-python
numpy

# ==============================
# DATA PROCESSING
# ==============================

pandas
scikit-learn

# ==============================
# OPTIONAL AI / NLP
# ==============================

spacy
transformers
sentence-transformers

# ==============================
# UTILITIES
# ==============================

tqdm
loguru

# ==============================
# TESTING
# ==============================

pytest



python-dotenv